changed array from short ot float

+++ app/src/main/java/com/degree/eliif/hearinganalyzer/POJO/Calibration.kt
@@ -3,12 +3,12 @@ package com.degree.eliif.hearinganalyzer.POJO
 import java.io.Serializable
 
 class Calibration : Serializable {
-  var calibrationLeft: MutableMap<Double, Double> = mutableMapOf()
-  var calibrationRight: MutableMap<Double, Double> = mutableMapOf()
+  var calibrationLeft: MutableMap<Double, Float> = mutableMapOf()
+  var calibrationRight: MutableMap<Double, Float> = mutableMapOf()
 
-  var dbHl: MutableMap<Double, Double> = mutableMapOf()
+  var dbHl: MutableMap<Double, Float> = mutableMapOf()
 
-  var measuringLevel: Double = 0.0
+  var measuringLevel: Float = 0.0F
 
   override fun toString(): String {
     return calibrationLeft.toString() + calibrationLeft.toString() + dbHl.toString() + measuringLevel.toString()
